pyproject.toml
[project]
name = "anvil"
version = "0.1.0"
description = "AI coding agent with structured tool support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "flake8>=7.0",
]
fetch = [
    "praw>=7.0",
    "playwright>=1.41",
]
gui = [
    "gradio>=4.0",
]
search = [
    "tavily-python>=0.3.5",
]

[project.scripts]
anvil = "anvil.cli:main"

[project.entry-points."scout.sources"]
hackernews = "scout.sources.hackernews:HackerNewsSource"
reddit = "scout.sources.reddit:RedditSource"
producthunt = "scout.sources.producthunt:ProductHuntSource"
github_issues = "scout.sources.github_issues:GitHubIssuesSource"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common", "src/anvil", "src/scout"]

[tool.pytest.ini_options]
testpaths = ["tests"]
